# api/openapi.yaml
openapi: 3.0.0
info:
  title: term-keeper-api
  version: "1.0.0"
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /terms:
    post:
      operationId: createTerm
      summary: Create a new term
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TermCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TermResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getTerms
      summary: Get a list of terms
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: false
          description: Query string for searching terms
          schema:
            type: string
        - name: category
          in: query
          required: false
          description: Category of the term
          schema:
            type: string
        - name: sort
          in: query
          required: false
          description: Sort order for the terms
          schema:
            type: string
            enum:
              - created_at_asc
              - created_at_desc
              - updated_at_asc
              - updated_at_desc
              - term_asc
              - term_desc
        - name: checked
          in: query
          required: false
          description: Filter by checked status
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TermListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TermCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryIds:
          type: array
          items:
            type: string
      required:
        - name
...
    TermResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryResponse"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TermListResponse:
      type: array
      items:
        $ref: "#/components/schemas/TermResponse"
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        hex_color_code:
          type: string
          pattern: "^#([0-9a-fA-F]{6}|[0-9a-fA-F]{3})$"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message